Provider2: Add package schemas
