+++ provider2/schemas/form_input/voids.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel, Field
+
+
+class VoidRequest(BaseModel):
+    amount: int = Field(..., ge=0, description="Amount must be a positive integer")
