+++ provider2/schemas/responses/transactions.py
@@ -11,15 +11,3 @@ class TransactionResponse(BaseModel):
     currency: str
     statementDescriptor: str
     cardId: uuid.UUID
-
-
-class VoidResponse(BaseModel):
-    id: uuid.UUID
-    created_at: str
-    status: str
-    original_amount: int
-    current_amount: int
-    currency: str
-    description: str
-    payment_method: str
-    card_id: uuid.UUID
